Add float16 data type for WebGPU (#6)
This is not available in WebAssembly yet! But it works in WebGPU where the `shader-f16` feature is supported. https://web3dsurvey.com/webgpu/features/shader-f16

I needed to upgrade Playwright to get Chrome 135+, which added Float16Array.

+++ src/numpy.ts
@@ -39,7 +39,7 @@ export const float32 = DType.Float32;
 export const int32 = DType.Int32;
 export const uint32 = DType.Uint32;
 export const bool = DType.Bool;
-export const complex64 = DType.Complex64;
+export const float16 = DType.Float16;
 
 // Constants section
